feat: refactor some functions for reuse and add a simpler script for fetching apy only
