+++ packages/scripts/process_yearn_vision/utils/yearn.py
@@ -1,11 +1,12 @@
 from decimal import Decimal
 from typing import Optional
 
+from web3 import exceptions
+
 from helpers.constants import Network
 from helpers.network import client, parse_json
 from helpers.web3 import Web3Provider
 from process_yearn_vision.typings import Address, Block, Vault
-from web3 import exceptions
 
 
 def get_vault(vault_address: Address, network: Network) -> Optional[Vault]:
